var() functions in shorthands: use a single code path with IDs
… rather than generating similar code for each shorthand.
